refactor: add attachments settings section for enhanced file management
- Introduced a new AttachmentsSettingsSection to manage attachment-related settings, allowing users to configure options for moving attachments with notes.
- Updated the AdvancedNoteMoverSettingsTab to include the new attachments settings, ensuring a cohesive user experience.
- Enhanced the settings to allow toggling of features such as moving attachments, skipping shared attachments, and deleting empty asset folders after moving.
- Refactored existing code to streamline attachment management functionality.
